requirements.txt
requests>=2.31.0
pandas>=2.0.0
jinja2>=3.1.0
yfinance>=0.2.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
